Add Pretter back in, along with lint-staged
The back-and-forth is killing me. However, this setup is designed to
represent as many of the best practices as possible along with the
general setup that Pendo itself uses.

We may in the future update the pre-commit hooks to run autofixes too,
but for now having the explicit console errors helps identify/debug
VSCode's format-on-save capabilities and incongruencies.